old/requirements.txt
requests
Pillow
python-dotenv
moviepy
gTTS
